fix warnings which are reported on CentOS 5.4/x86_64
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@4416 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ gftool/gfwhere/gfwhere.c
@@ -54,7 +54,7 @@ display_replica(struct gfs_stat *st, struct gfs_replica_info *ri, int i,
 	if (opt->type_suffix) {
 		if (opt->print_dead_copy &&
 		    (igen = gfs_replica_info_nth_gen(ri, i)) != st->st_gen)
-			printf(";%llu", igen);
+			printf(";%llu", (unsigned long long)igen);
 		if (gfs_replica_info_nth_is_incomplete(ri, i))
 			putchar('?');
 		if (gfs_replica_info_nth_is_dead_host(ri, i))
@@ -67,7 +67,8 @@ display_replica(struct gfs_stat *st, struct gfs_replica_info *ri, int i,
 		    opt->print_dead_copy)
 			putchar('\t');
 		if (opt->print_dead_copy)
-			printf("%10llu\t", gfs_replica_info_nth_gen(ri, i));
+			printf("%10llu\t", (unsigned long long)
+			    gfs_replica_info_nth_gen(ri, i));
 		need_space = 0;
 		if (opt->print_dead_host) {
 			putchar(gfs_replica_info_nth_is_dead_host(ri, i) ?
